renamed file to match other enums standard
