+++ lib/gfx/gfx.js
@@ -27,7 +27,7 @@ const SpritesGroup = require('./sprites_group.js');
 const Vector2 = require('../utils/vector2.js');
 const FontTextureAsset = require('../assets/font_texture_asset.js');
 const MsdfFontTextureAsset = require('../assets/msdf_font_texture_asset.js');
-const { TextAlignment, TextAlignments } = require('./text_alignment.js');
+const { TextAlignment, TextAlignments } = require('./text_alignments.js');
 const Mesh = require('./mesh.js');
 const Circle = require('../utils/circle.js');
 const SpriteBatch = require('./sprite_batch.js');
