+++ lib/gfx/text_alignments.js
